+++ secrets/consul/backend_test.go
@@ -5,12 +5,10 @@ package consul
 
 import (
 	"context"
-	"encoding/base64"
 	"fmt"
 	"log"
 	"os"
 	"reflect"
-	"strings"
 	"testing"
 	"time"
 
@@ -21,24 +19,32 @@ import (
 	"github.com/openbao/openbao/sdk/v2/logical"
 )
 
+func testOldestAndLatestSupported(t *testing.T, f func(t *testing.T, versionId string)) {
+	t.Run("latest-supported", func(t *testing.T) {
+		f(t, "latest-supported")
+	})
+
+	t.Run("oldest-supported", func(t *testing.T) {
+		f(t, "oldest-supported")
+	})
+}
+
 func TestBackend_Config_Access(t *testing.T) {
-	t.Run("config_access", func(t *testing.T) {
+	t.Parallel()
+
+	t.Run("no automatic bootstrap", func(t *testing.T) {
 		t.Parallel()
-		t.Run("pre-1.4.0", func(t *testing.T) {
+		testOldestAndLatestSupported(t, func(t *testing.T, versionId string) {
 			t.Parallel()
-			testBackendConfigAccess(t, "1.3.1", true)
+			testBackendConfigAccess(t, versionId, false)
 		})
-		t.Run("post-1.4.0", func(t *testing.T) {
+	})
+
+	t.Run("automatic bootstrap", func(t *testing.T) {
+		t.Parallel()
+		testOldestAndLatestSupported(t, func(t *testing.T, versionId string) {
 			t.Parallel()
-			testBackendConfigAccess(t, "", true)
-		})
-		t.Run("pre-1.4.0 automatic-bootstrap", func(t *testing.T) {
-			t.Parallel()
-			testBackendConfigAccess(t, "1.3.1", false)
-		})
-		t.Run("post-1.4.0 automatic-bootstrap", func(t *testing.T) {
-			t.Parallel()
-			testBackendConfigAccess(t, "", false)
+			testBackendConfigAccess(t, versionId, true)
 		})
 	})
 }
@@ -51,13 +57,17 @@ func testBackendConfigAccess(t *testing.T, version string, autoBootstrap bool) {
 		t.Fatal(err)
 	}
 
-	cleanup, consulConfig := consul.PrepareTestContainer(t, version, false, autoBootstrap)
+	cleanup, consulConfig := consul.PrepareTestContainer(t, version, false, !autoBootstrap)
 	defer cleanup()
 
-	connData := map[string]interface{}{
+	connData := map[string]any{
 		"address": consulConfig.Address(),
 	}
-	if autoBootstrap || strings.HasPrefix(version, "1.3") {
+	if autoBootstrap {
+		if consulConfig.Token != "" {
+			t.Fatal("expected bootstrap not to have happened yet")
+		}
+	} else {
 		connData["token"] = consulConfig.Token
 	}
 
@@ -79,7 +89,7 @@ func testBackendConfigAccess(t *testing.T, version string, autoBootstrap bool) {
 		t.Fatalf("failed to write configuration: resp:%#v err:%s", resp, err)
 	}
 
-	expected := map[string]interface{}{
+	expected := map[string]any{
 		"address": connData["address"].(string),
 		"scheme":  "http",
 	}
@@ -92,32 +102,10 @@ func testBackendConfigAccess(t *testing.T, version string, autoBootstrap bool) {
 }
 
 func TestBackend_Renew_Revoke(t *testing.T) {
-	t.Run("renew_revoke", func(t *testing.T) {
+	t.Parallel()
+	testOldestAndLatestSupported(t, func(t *testing.T, versionId string) {
 		t.Parallel()
-		t.Run("pre-1.4.0", func(t *testing.T) {
-			t.Parallel()
-			testBackendRenewRevoke(t, "1.3.1")
-		})
-		t.Run("post-1.4.0", func(t *testing.T) {
-			t.Parallel()
-			t.Run("legacy", func(t *testing.T) {
-				t.Parallel()
-				testBackendRenewRevoke(t, "1.4.4")
-			})
-
-			t.Run("param-policies", func(t *testing.T) {
-				t.Parallel()
-				testBackendRenewRevoke14(t, "", "policies")
-			})
-			t.Run("param-consul_policies", func(t *testing.T) {
-				t.Parallel()
-				testBackendRenewRevoke14(t, "", "consul_policies")
-			})
-			t.Run("both-params", func(t *testing.T) {
-				t.Parallel()
-				testBackendRenewRevoke14(t, "", "both")
-			})
-		})
+		testBackendRenewRevoke(t, versionId)
 	})
 }
 
@@ -132,7 +120,7 @@ func testBackendRenewRevoke(t *testing.T, version string) {
 	cleanup, consulConfig := consul.PrepareTestContainer(t, version, false, true)
 	defer cleanup()
 
-	connData := map[string]interface{}{
+	connData := map[string]any{
 		"address": consulConfig.Address(),
 		"token":   consulConfig.Token,
 	}
@@ -149,9 +137,9 @@ func testBackendRenewRevoke(t *testing.T, version string) {
 	}
 
 	req.Path = "roles/test"
-	req.Data = map[string]interface{}{
-		"policy": base64.StdEncoding.EncodeToString([]byte(testPolicy)),
-		"lease":  "6h",
+	req.Data = map[string]any{
+		"consul_policies": []string{"test"},
+		"lease":           "6h",
 	}
 	_, err = b.HandleRequest(context.Background(), req)
 	if err != nil {
@@ -234,7 +222,7 @@ func testBackendRenewRevoke14(t *testing.T, version string, policiesParam string
 	cleanup, consulConfig := consul.PrepareTestContainer(t, version, false, true)
 	defer cleanup()
 
-	connData := map[string]interface{}{
+	connData := map[string]any{
 		"address": consulConfig.Address(),
 		"token":   consulConfig.Token,
 	}
@@ -251,7 +239,7 @@ func testBackendRenewRevoke14(t *testing.T, version string, policiesParam string
 	}
 
 	req.Path = "roles/test"
-	req.Data = map[string]interface{}{
+	req.Data = map[string]any{
 		"lease": "6h",
 	}
 	if policiesParam == "both" {
@@ -264,25 +252,6 @@ func testBackendRenewRevoke14(t *testing.T, version string, policiesParam string
 	_, err = b.HandleRequest(context.Background(), req)
 	if err != nil {
 		t.Fatal(err)
-	}
-
-	read := &logical.Request{
-		Storage:   config.StorageView,
-		Operation: logical.ReadOperation,
-		Path:      "roles/test",
-		Data:      connData,
-	}
-	roleResp, err := b.HandleRequest(context.Background(), read)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	expectExtract := roleResp.Data["consul_policies"]
-	respExtract := roleResp.Data[policiesParam]
-	if respExtract != nil {
-		if expectExtract.([]string)[0] != respExtract.([]string)[0] {
-			t.Errorf("mismatch: response consul_policies '%s' does not match '[test]'", roleResp.Data["consul_policies"])
-		}
 	}
 
 	req.Operation = logical.ReadOperation
@@ -411,10 +380,10 @@ func TestBackend_LocalToken(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	cleanup, consulConfig := consul.PrepareTestContainer(t, "", false, true)
+	cleanup, consulConfig := consul.PrepareTestContainer(t, "latest-supported", false, true)
 	defer cleanup()
 
-	connData := map[string]interface{}{
+	connData := map[string]any{
 		"address": consulConfig.Address(),
 		"token":   consulConfig.Token,
 	}
@@ -431,7 +400,7 @@ func TestBackend_LocalToken(t *testing.T) {
 	}
 
 	req.Path = "roles/test"
-	req.Data = map[string]interface{}{
+	req.Data = map[string]any{
 		"consul_policies": []string{"test"},
 		"ttl":             "6h",
 		"local":           false,
@@ -442,7 +411,7 @@ func TestBackend_LocalToken(t *testing.T) {
 	}
 
 	req.Path = "roles/test_local"
-	req.Data = map[string]interface{}{
+	req.Data = map[string]any{
 		"consul_policies": []string{"test"},
 		"ttl":             "6h",
 		"local":           true,
@@ -528,64 +497,11 @@ func TestBackend_LocalToken(t *testing.T) {
 	}
 }
 
-func TestBackend_Management(t *testing.T) {
-	t.Run("management", func(t *testing.T) {
-		t.Parallel()
-		t.Run("pre-1.4.0", func(t *testing.T) {
-			t.Parallel()
-			testBackendManagement(t, "1.3.1")
-		})
-		t.Run("post-1.4.0", func(t *testing.T) {
-			t.Parallel()
-			testBackendManagement(t, "1.4.4")
-		})
-
-		testBackendManagement(t, "1.10.8")
-	})
-}
-
-func testBackendManagement(t *testing.T, version string) {
-	config := logical.TestBackendConfig()
-	config.StorageView = &logical.InmemStorage{}
-	b, err := Factory(context.Background(), config)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	cleanup, consulConfig := consul.PrepareTestContainer(t, version, false, true)
-	defer cleanup()
-
-	connData := map[string]interface{}{
-		"address": consulConfig.Address(),
-		"token":   consulConfig.Token,
-	}
-
-	logicaltest.Test(t, logicaltest.TestCase{
-		LogicalBackend: b,
-		Steps: []logicaltest.TestStep{
-			testAccStepConfig(t, connData),
-			testAccStepWriteManagementPolicy(t, "test", ""),
-			testAccStepReadManagementToken(t, "test", connData),
-		},
-	})
-}
-
 func TestBackend_Basic(t *testing.T) {
-	t.Run("basic", func(t *testing.T) {
+	t.Parallel()
+	testOldestAndLatestSupported(t, func(t *testing.T, versionId string) {
 		t.Parallel()
-		t.Run("pre-1.4.0", func(t *testing.T) {
-			t.Parallel()
-			testBackendBasic(t, "1.3.1")
-		})
-		t.Run("post-1.4.0", func(t *testing.T) {
-			t.Parallel()
-			t.Run("legacy", func(t *testing.T) {
-				t.Parallel()
-				testBackendBasic(t, "1.4.4")
-			})
-
-			testBackendBasic(t, "1.10.8")
-		})
+		testBackendBasic(t, versionId)
 	})
 }
 
@@ -600,7 +516,7 @@ func testBackendBasic(t *testing.T, version string) {
 	cleanup, consulConfig := consul.PrepareTestContainer(t, version, false, true)
 	defer cleanup()
 
-	connData := map[string]interface{}{
+	connData := map[string]any{
 		"address": consulConfig.Address(),
 		"token":   consulConfig.Token,
 	}
@@ -609,7 +525,7 @@ func testBackendBasic(t *testing.T, version string) {
 		LogicalBackend: b,
 		Steps: []logicaltest.TestStep{
 			testAccStepConfig(t, connData),
-			testAccStepWritePolicy(t, "test", testPolicy, ""),
+			testAccStepWriteRole(t, "test", "test", ""),
 			testAccStepReadToken(t, "test", connData),
 		},
 	})
@@ -620,29 +536,29 @@ func TestBackend_crud(t *testing.T) {
 	logicaltest.Test(t, logicaltest.TestCase{
 		LogicalBackend: b,
 		Steps: []logicaltest.TestStep{
-			testAccStepWritePolicy(t, "test", testPolicy, ""),
-			testAccStepWritePolicy(t, "test2", testPolicy, ""),
-			testAccStepWritePolicy(t, "test3", testPolicy, ""),
-			testAccStepReadPolicy(t, "test", testPolicy, 0),
-			testAccStepListPolicy(t, []string{"test", "test2", "test3"}),
-			testAccStepDeletePolicy(t, "test"),
+			testAccStepWriteRole(t, "test", "write", ""),
+			testAccStepWriteRole(t, "test2", "write", ""),
+			testAccStepWriteRole(t, "test3", "write", ""),
+			testAccStepReadRole(t, "test", "write", 0),
+			testAccStepListRole(t, []string{"test", "test2", "test3"}),
+			testAccStepDeleteRole(t, "test"),
 		},
 	})
 }
 
-func TestBackend_role_lease(t *testing.T) {
+func TestBackend_role_ttl(t *testing.T) {
 	b, _ := Factory(context.Background(), logical.TestBackendConfig())
 	logicaltest.Test(t, logicaltest.TestCase{
 		LogicalBackend: b,
 		Steps: []logicaltest.TestStep{
-			testAccStepWritePolicy(t, "test", testPolicy, "6h"),
-			testAccStepReadPolicy(t, "test", testPolicy, 6*time.Hour),
-			testAccStepDeletePolicy(t, "test"),
+			testAccStepWriteRole(t, "test", "write", "6h"),
+			testAccStepReadRole(t, "test", "write", 6*time.Hour),
+			testAccStepDeleteRole(t, "test"),
 		},
 	})
 }
 
-func testAccStepConfig(t *testing.T, config map[string]interface{}) logicaltest.TestStep {
+func testAccStepConfig(t *testing.T, config map[string]any) logicaltest.TestStep {
 	return logicaltest.TestStep{
 		Operation: logical.UpdateOperation,
 		Path:      "config/access",
@@ -650,7 +566,7 @@ func testAccStepConfig(t *testing.T, config map[string]interface{}) logicaltest.
 	}
 }
 
-func testAccStepReadToken(t *testing.T, name string, conf map[string]interface{}) logicaltest.TestStep {
+func testAccStepReadToken(t *testing.T, name string, conf map[string]any) logicaltest.TestStep {
 	return logicaltest.TestStep{
 		Operation: logical.ReadOperation,
 		Path:      "creds/" + name,
@@ -686,88 +602,37 @@ func testAccStepReadToken(t *testing.T, name string, conf map[string]interface{}
 	}
 }
 
-func testAccStepReadManagementToken(t *testing.T, name string, conf map[string]interface{}) logicaltest.TestStep {
-	return logicaltest.TestStep{
-		Operation: logical.ReadOperation,
-		Path:      "creds/" + name,
-		Check: func(resp *logical.Response) error {
-			var d struct {
-				Token string `mapstructure:"token"`
-			}
-			if err := mapstructure.Decode(resp.Data, &d); err != nil {
-				return err
-			}
-			log.Printf("[WARN] Generated token: %s", d.Token)
-
-			// Build a client and verify that the credentials work
-			config := consulapi.DefaultConfig()
-			config.Address = conf["address"].(string)
-			config.Token = d.Token
-			client, err := consulapi.NewClient(config)
-			if err != nil {
-				return err
-			}
-
-			log.Printf("[WARN] Verifying that the generated token works...")
-			_, _, err = client.ACL().Create(&consulapi.ACLEntry{ //nolint:staticcheck
-				Type: "management",
-				Name: "test2",
-			}, nil)
-			if err != nil {
-				return err
-			}
-
-			return nil
-		},
-	}
-}
-
-func testAccStepWritePolicy(t *testing.T, name string, policy string, lease string) logicaltest.TestStep {
+func testAccStepWriteRole(t *testing.T, name string, policy string, ttl string) logicaltest.TestStep {
 	return logicaltest.TestStep{
 		Operation: logical.UpdateOperation,
 		Path:      "roles/" + name,
-		Data: map[string]interface{}{
-			"policy": base64.StdEncoding.EncodeToString([]byte(policy)),
-			"lease":  lease,
+		Data: map[string]any{
+			"consul_policies": []string{policy},
+			"ttl":             ttl,
 		},
 	}
 }
 
-func testAccStepWriteManagementPolicy(t *testing.T, name string, lease string) logicaltest.TestStep {
-	return logicaltest.TestStep{
-		Operation: logical.UpdateOperation,
-		Path:      "roles/" + name,
-		Data: map[string]interface{}{
-			"token_type": "management",
-			"lease":      lease,
-		},
-	}
-}
-
-func testAccStepReadPolicy(t *testing.T, name string, policy string, lease time.Duration) logicaltest.TestStep {
+func testAccStepReadRole(t *testing.T, name string, policy string, ttl time.Duration) logicaltest.TestStep {
 	return logicaltest.TestStep{
 		Operation: logical.ReadOperation,
 		Path:      "roles/" + name,
 		Check: func(resp *logical.Response) error {
-			policyRaw := resp.Data["policy"].(string)
-			out, err := base64.StdEncoding.DecodeString(policyRaw)
-			if err != nil {
-				return err
-			}
-			if string(out) != policy {
-				return fmt.Errorf("mismatch: %s %s", out, policy)
+			policies := resp.Data["consul_policies"].([]string)
+			if len(policies) != 1 || policies[0] != policy {
+				return fmt.Errorf("mismatch: %q %v", policy, policies)
 			}
 
-			l := resp.Data["lease"].(int64)
-			if lease != time.Second*time.Duration(l) {
-				return fmt.Errorf("mismatch: %v %v", l, lease)
+			l := resp.Data["ttl"].(int64)
+			if ttl != time.Second*time.Duration(l) {
+				return fmt.Errorf("mismatch: %v %v", l, ttl)
 			}
 			return nil
 		},
 	}
 }
 
-func testAccStepListPolicy(t *testing.T, names []string) logicaltest.TestStep {
+func testAccStepListRole(t *testing.T, names []string) logicaltest.TestStep {
 	return logicaltest.TestStep{
 		Operation: logical.ListOperation,
 		Path:      "roles/",
@@ -781,7 +646,7 @@ func testAccStepListPolicy(t *testing.T, names []string) logicaltest.TestStep {
 	}
 }
 
-func testAccStepDeletePolicy(t *testing.T, name string) logicaltest.TestStep {
+func testAccStepDeleteRole(t *testing.T, name string) logicaltest.TestStep {
 	return logicaltest.TestStep{
 		Operation: logical.DeleteOperation,
 		Path:      "roles/" + name,
@@ -796,10 +661,10 @@ func TestBackend_Roles(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	cleanup, consulConfig := consul.PrepareTestContainer(t, "", false, true)
+	cleanup, consulConfig := consul.PrepareTestContainer(t, "latest-supported", false, true)
 	defer cleanup()
 
-	connData := map[string]interface{}{
+	connData := map[string]any{
 		"address": consulConfig.Address(),
 		"token":   consulConfig.Token,
 	}
@@ -817,9 +682,9 @@ func TestBackend_Roles(t *testing.T) {
 
 	// Create the consul_roles role
 	req.Path = "roles/test-consul-roles"
-	req.Data = map[string]interface{}{
+	req.Data = map[string]any{
 		"consul_roles": []string{"role-test"},
-		"lease":        "6h",
+		"ttl":          "6h",
 	}
 	_, err = b.HandleRequest(context.Background(), req)
 	if err != nil {
@@ -938,7 +803,7 @@ func testBackendEntDiffNamespaceRevocation(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	cleanup, consulConfig := consul.PrepareTestContainer(t, "", true, true)
+	cleanup, consulConfig := consul.PrepareTestContainer(t, "latest-supported", true, true)
 	defer cleanup()
 
 	// Perform additional Consul configuration
@@ -975,7 +840,7 @@ func testBackendEntDiffNamespaceRevocation(t *testing.T) {
 	}
 
 	// Write backend config
-	connData := map[string]interface{}{
+	connData := map[string]any{
 		"address": consulConfig.Address(),
 		"token":   cToken.SecretID,
 	}
@@ -993,9 +858,9 @@ func testBackendEntDiffNamespaceRevocation(t *testing.T) {
 
 	// Create the role in namespace "ns1"
 	req.Path = "roles/test-ns"
-	req.Data = map[string]interface{}{
+	req.Data = map[string]any{
 		"consul_policies":  []string{"ns-test"},
-		"lease":            "6h",
+		"ttl":              "6h",
 		"consul_namespace": "ns1",
 	}
 	_, err = b.HandleRequest(context.Background(), req)
@@ -1069,7 +934,7 @@ func testBackendEntDiffPartitionRevocation(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	cleanup, consulConfig := consul.PrepareTestContainer(t, "", true, true)
+	cleanup, consulConfig := consul.PrepareTestContainer(t, "latest-supported", true, true)
 	defer cleanup()
 
 	// Perform additional Consul configuration
@@ -1106,7 +971,7 @@ func testBackendEntDiffPartitionRevocation(t *testing.T) {
 	}
 
 	// Write backend config
-	connData := map[string]interface{}{
+	connData := map[string]any{
 		"address": consulConfig.Address(),
 		"token":   cToken.SecretID,
 	}
@@ -1124,9 +989,9 @@ func testBackendEntDiffPartitionRevocation(t *testing.T) {
 
 	// Create the role in partition "part1"
 	req.Path = "roles/test-part"
-	req.Data = map[string]interface{}{
+	req.Data = map[string]any{
 		"consul_policies": []string{"part-test"},
-		"lease":           "6h",
+		"ttl":             "6h",
 		"partition":       "part1",
 	}
 	_, err = b.HandleRequest(context.Background(), req)
@@ -1200,10 +1065,10 @@ func testBackendEntNamespace(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	cleanup, consulConfig := consul.PrepareTestContainer(t, "", true, true)
+	cleanup, consulConfig := consul.PrepareTestContainer(t, "latest-supported", true, true)
 	defer cleanup()
 
-	connData := map[string]interface{}{
+	connData := map[string]any{
 		"address": consulConfig.Address(),
 		"token":   consulConfig.Token,
 	}
@@ -1221,9 +1086,9 @@ func testBackendEntNamespace(t *testing.T) {
 
 	// Create the role in namespace "ns1"
 	req.Path = "roles/test-ns"
-	req.Data = map[string]interface{}{
+	req.Data = map[string]any{
 		"consul_policies":  []string{"ns-test"},
-		"lease":            "6h",
+		"ttl":              "6h",
 		"consul_namespace": "ns1",
 	}
 	_, err = b.HandleRequest(context.Background(), req)
@@ -1317,10 +1182,10 @@ func testBackendEntPartition(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	cleanup, consulConfig := consul.PrepareTestContainer(t, "", true, true)
+	cleanup, consulConfig := consul.PrepareTestContainer(t, "latest-supported", true, true)
 	defer cleanup()
 
-	connData := map[string]interface{}{
+	connData := map[string]any{
 		"address": consulConfig.Address(),
 		"token":   consulConfig.Token,
 	}
@@ -1338,9 +1203,9 @@ func testBackendEntPartition(t *testing.T) {
 
 	// Create the role in partition "part1"
 	req.Path = "roles/test-part"
-	req.Data = map[string]interface{}{
+	req.Data = map[string]any{
 		"consul_policies": []string{"part-test"},
-		"lease":           "6h",
+		"ttl":             "6h",
 		"partition":       "part1",
 	}
 	_, err = b.HandleRequest(context.Background(), req)
@@ -1434,10 +1299,10 @@ func TestBackendRenewRevokeRolesAndIdentities(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	cleanup, consulConfig := consul.PrepareTestContainer(t, "", false, true)
+	cleanup, consulConfig := consul.PrepareTestContainer(t, "latest-supported", false, true)
 	defer cleanup()
 
-	connData := map[string]interface{}{
+	connData := map[string]any{
 		"address": consulConfig.Address(),
 		"token":   consulConfig.Token,
 	}
@@ -1455,121 +1320,121 @@ func TestBackendRenewRevokeRolesAndIdentities(t *testing.T) {
 
 	cases := map[string]struct {
 		RoleName string
-		RoleData map[string]interface{}
+		RoleData map[string]any
 	}{
 		"just role": {
 			"r",
-			map[string]interface{}{
+			map[string]any{
 				"consul_roles": []string{"role-test"},
-				"lease":        "6h",
+				"ttl":          "6h",
 			},
 		},
 		"role and policies": {
 			"rp",
-			map[string]interface{}{
+			map[string]any{
 				"consul_policies": []string{"test"},
 				"consul_roles":    []string{"role-test"},
-				"lease":           "6h",
+				"ttl":             "6h",
 			},
 		},
 		"service identity": {
 			"si",
-			map[string]interface{}{
+			map[string]any{
 				"service_identities": "service1",
-				"lease":              "6h",
+				"ttl":                "6h",
 			},
 		},
 		"service identity and policies": {
 			"sip",
-			map[string]interface{}{
+			map[string]any{
 				"consul_policies":    []string{"test"},
 				"service_identities": "service1",
-				"lease":              "6h",
+				"ttl":                "6h",
 			},
 		},
 		"service identity and role": {
 			"sir",
-			map[string]interface{}{
+			map[string]any{
 				"consul_roles":       []string{"role-test"},
 				"service_identities": "service1",
-				"lease":              "6h",
+				"ttl":                "6h",
 			},
 		},
 		"service identity and role and policies": {
 			"sirp",
-			map[string]interface{}{
+			map[string]any{
 				"consul_policies":    []string{"test"},
 				"consul_roles":       []string{"role-test"},
 				"service_identities": "service1",
-				"lease":              "6h",
+				"ttl":                "6h",
 			},
 		},
 		"node identity": {
 			"ni",
-			map[string]interface{}{
+			map[string]any{
 				"node_identities": []string{"node1:dc1"},
-				"lease":           "6h",
+				"ttl":             "6h",
 			},
 		},
 		"node identity and policies": {
 			"nip",
-			map[string]interface{}{
+			map[string]any{
 				"consul_policies": []string{"test"},
 				"node_identities": []string{"node1:dc1"},
-				"lease":           "6h",
+				"ttl":             "6h",
 			},
 		},
 		"node identity and role": {
 			"nir",
-			map[string]interface{}{
+			map[string]any{
 				"consul_roles":    []string{"role-test"},
 				"node_identities": []string{"node1:dc1"},
-				"lease":           "6h",
+				"ttl":             "6h",
 			},
 		},
 		"node identity and role and policies": {
 			"nirp",
-			map[string]interface{}{
+			map[string]any{
 				"consul_policies": []string{"test"},
 				"consul_roles":    []string{"role-test"},
 				"node_identities": []string{"node1:dc1"},
-				"lease":           "6h",
+				"ttl":             "6h",
 			},
 		},
 		"node identity and service identity": {
 			"nisi",
-			map[string]interface{}{
+			map[string]any{
 				"service_identities": "service1",
 				"node_identities":    []string{"node1:dc1"},
-				"lease":              "6h",
+				"ttl":                "6h",
 			},
 		},
 		"node identity and service identity and policies": {
 			"nisip",
-			map[string]interface{}{
+			map[string]any{
 				"consul_policies":    []string{"test"},
 				"service_identities": "service1",
 				"node_identities":    []string{"node1:dc1"},
-				"lease":              "6h",
+				"ttl":                "6h",
 			},
 		},
 		"node identity and service identity and role": {
 			"nisir",
-			map[string]interface{}{
+			map[string]any{
 				"consul_roles":       []string{"role-test"},
 				"service_identities": "service1",
 				"node_identities":    []string{"node1:dc1"},
-				"lease":              "6h",
+				"ttl":                "6h",
 			},
 		},
 		"node identity and service identity and role and policies": {
 			"nisirp",
-			map[string]interface{}{
+			map[string]any{
 				"consul_policies":    []string{"test"},
 				"consul_roles":       []string{"role-test"},
 				"service_identities": "service1",
 				"node_identities":    []string{"node1:dc1"},
-				"lease":              "6h",
+				"ttl":                "6h",
 			},
 		},
 	}
@@ -1658,8 +1523,3 @@ func TestBackendRenewRevokeRolesAndIdentities(t *testing.T) {
 		}
 	}
 }
-
-const testPolicy = `
-key "" {
-	policy = "write"
-}`