+++ secrets/consul/path_roles.go
@@ -5,8 +5,6 @@ package consul
 
 import (
 	"context"
-	"encoding/base64"
-	"fmt"
 	"time"
 
 	"github.com/openbao/openbao/sdk/v2/framework"
@@ -43,30 +41,6 @@ func pathRoles(b *backend) *framework.Path {
 				Description: "Name of the role.",
 			},
 
-			// The "policy" and "token_type" parameters were deprecated in Consul back in version 1.4.
-			// They have been removed from Consul as of version 1.11. Consider removing them here in the future.
-			"policy": {
-				Type: framework.TypeString,
-				Description: `Policy document, base64 encoded. Required
-for 'client' tokens. Required for Consul pre-1.4.`,
-				Deprecated: true,
-			},
-
-			"token_type": {
-				Type:    framework.TypeString,
-				Default: "client",
-				Description: `Which type of token to create: 'client' or 'management'. If
-a 'management' token, the "policy", "policies", and "consul_roles" parameters are not
-required. Defaults to 'client'.`,
-				Deprecated: true,
-			},
-
-			"policies": {
-				Type:        framework.TypeCommaStringSlice,
-				Description: `Use "consul_policies" instead.`,
-				Deprecated:  true,
-			},
-
 			"consul_policies": {
 				Type: framework.TypeCommaStringSlice,
 				Description: `List of policies to attach to the token. Either "consul_policies"
@@ -76,14 +50,14 @@ using Consul 1.4.`,
 
 			"consul_roles": {
 				Type: framework.TypeCommaStringSlice,
-				Description: `List of Consul roles to attach to the token. Either "policies"
+				Description: `List of Consul roles to attach to the token. Either "consul_policies"
 or "consul_roles" are required for Consul 1.5 and above.`,
 			},
 
 			"local": {
 				Type: framework.TypeBool,
 				Description: `Indicates that the token should not be replicated globally 
-and instead be local to the current datacenter. Available in Consul 1.4 and above.`,
+and instead be local to the current datacenter.`,
 			},
 
 			"ttl": {
@@ -94,12 +68,6 @@ and instead be local to the current datacenter. Available in Consul 1.4 and abov
 			"max_ttl": {
 				Type:        framework.TypeDurationSecond,
 				Description: "Max TTL for the Consul token created from the role.",
-			},
-
-			"lease": {
-				Type:        framework.TypeDurationSecond,
-				Description: `Use "ttl" instead.`,
-				Deprecated:  true,
 			},
 
 			"consul_namespace": {
@@ -160,24 +128,15 @@ func (b *backend) pathRolesRead(ctx context.Context, req *logical.Request, d *fr
 		return nil, err
 	}
 
-	if roleConfigData.TokenType == "" {
-		roleConfigData.TokenType = "client"
-	}
-
 	// Generate the response
 	resp := &logical.Response{
-		Data: map[string]interface{}{
-			"lease":            int64(roleConfigData.TTL.Seconds()),
+		Data: map[string]any{
 			"ttl":              int64(roleConfigData.TTL.Seconds()),
 			"max_ttl":          int64(roleConfigData.MaxTTL.Seconds()),
-			"token_type":       roleConfigData.TokenType,
 			"local":            roleConfigData.Local,
 			"consul_namespace": roleConfigData.ConsulNamespace,
 			"partition":        roleConfigData.Partition,
 		},
-	}
-	if roleConfigData.Policy != "" {
-		resp.Data["policy"] = base64.StdEncoding.EncodeToString([]byte(roleConfigData.Policy))
 	}
 	if len(roleConfigData.Policies) > 0 {
 		resp.Data["consul_policies"] = roleConfigData.Policies
@@ -196,45 +155,15 @@ func (b *backend) pathRolesRead(ctx context.Context, req *logical.Request, d *fr
 }
 
 func (b *backend) pathRolesWrite(ctx context.Context, req *logical.Request, d *framework.FieldData) (*logical.Response, error) {
-	tokenType := d.Get("token_type").(string)
-	policy := d.Get("policy").(string)
 	consulPolicies := d.Get("consul_policies").([]string)
-	policies := d.Get("policies").([]string)
 	roles := d.Get("consul_roles").([]string)
 	serviceIdentities := d.Get("service_identities").([]string)
 	nodeIdentities := d.Get("node_identities").([]string)
-
-	switch tokenType {
-	case "client":
-		if policy == "" && len(policies) == 0 && len(consulPolicies) == 0 &&
-			len(roles) == 0 && len(serviceIdentities) == 0 && len(nodeIdentities) == 0 {
-			return logical.ErrorResponse(
-				"Use either a policy document, a list of policies or roles, or a set of service or node identities, depending on your Consul version"), nil
-		}
-	case "management":
-	default:
-		return logical.ErrorResponse("token_type must be \"client\" or \"management\""), nil
-	}
-
-	if len(consulPolicies) == 0 {
-		consulPolicies = policies
-	}
-
-	policyRaw, err := base64.StdEncoding.DecodeString(policy)
-	if err != nil {
-		return logical.ErrorResponse(fmt.Sprintf(
-			"Error decoding policy base64: %s", err)), nil
-	}
 
 	var ttl time.Duration
 	ttlRaw, ok := d.GetOk("ttl")
 	if ok {
 		ttl = time.Second * time.Duration(ttlRaw.(int))
-	} else {
-		leaseParamRaw, ok := d.GetOk("lease")
-		if ok {
-			ttl = time.Second * time.Duration(leaseParamRaw.(int))
-		}
 	}
 
 	var maxTTL time.Duration
@@ -248,12 +177,10 @@ func (b *backend) pathRolesWrite(ctx context.Context, req *logical.Request, d *f
 	namespace := d.Get("consul_namespace").(string)
 	partition := d.Get("partition").(string)
 	entry, err := logical.StorageEntryJSON("policy/"+name, roleConfig{
-		Policy:            string(policyRaw),
 		Policies:          consulPolicies,
 		ConsulRoles:       roles,
 		ServiceIdentities: serviceIdentities,
 		NodeIdentities:    nodeIdentities,
-		TokenType:         tokenType,
 		TTL:               ttl,
 		MaxTTL:            maxTTL,
 		Local:             local,
@@ -280,14 +207,12 @@ func (b *backend) pathRolesDelete(ctx context.Context, req *logical.Request, d *
 }
 
 type roleConfig struct {
-	Policy            string        `json:"policy"`
 	Policies          []string      `json:"policies"`
 	ConsulRoles       []string      `json:"consul_roles"`
 	ServiceIdentities []string      `json:"service_identities"`
 	NodeIdentities    []string      `json:"node_identities"`
 	TTL               time.Duration `json:"lease"`
 	MaxTTL            time.Duration `json:"max_ttl"`
-	TokenType         string        `json:"token_type"`
 	Local             bool          `json:"local"`
 	ConsulNamespace   string        `json:"consul_namespace"`
 	Partition         string        `json:"partition"`